fix(core): Migrate UserMenu / AccountMenu to NcListItem
